requirements.txt
# AI Research Agent Team — Python Dependencies
# Install with: pip install -r requirements.txt
# Python 3.10+ required

# ── Core LLM Frameworks ──────────────────────────────────────
langchain>=0.3.0
langchain-core>=0.3.0
langchain-openai>=0.2.0
langchain-anthropic>=0.2.0
langchain-google-genai>=2.0.0
langchain-ollama>=0.2.0               # Local models via Ollama
langchain-community>=0.3.0            # Tavily search tool + community integrations
langgraph>=0.2.0

# ── Multi-Agent Frameworks (install the one you use) ─────────
crewai>=0.80.0
pyautogen>=0.4.0

# ── Search Tools ─────────────────────────────────────────────
tavily-python>=0.5.0
google-search-results>=2.4.2          # Serper / SerpAPI

# ── Web Scraping ─────────────────────────────────────────────
firecrawl-py>=1.0.0
beautifulsoup4>=4.12.0
httpx>=0.27.0                         # Also used for Ollama / LM Studio connectivity detection
playwright>=1.45.0                    # For JS-heavy pages

# ── Data Processing ──────────────────────────────────────────
pydantic>=2.7.0
python-dotenv>=1.0.0
PyYAML>=6.0.1
jinja2>=3.1.4

# ── Email Delivery ───────────────────────────────────────────
aiosmtplib>=3.0.0                     # Async SMTP

# ── Optional Integrations ────────────────────────────────────
slack-sdk>=3.29.0                     # Slack delivery
notion-client>=2.2.2                  # Notion delivery

# ── Observability ────────────────────────────────────────────
langsmith>=0.1.0                      # LangGraph tracing

# ── Development & Testing ────────────────────────────────────
pytest>=8.2.0
pytest-asyncio>=0.23.0
rich>=13.7.0                          # Pretty CLI output
